Modify game loop for proper rounds to end game
I was unable to make the computer choose a new status at the beginning
of each round. This was coupled with scattered code and improper syntax.
The return value was also undefined, so that needed to be modified.

I cleaned up a bit of the code, modified bits to accept more values,
and reworked the game loop so that the player could win, lose, or tie
the game at random. The return value was also changed to indicate the
game was over. Syntax was fixed up a bit, by adding missing ;